feat: 내가 User거나 상대가 User가 아닐때 ban, mute 금지 프론트 처리
