--- frontend/src/components/commons/NicknameMenu/reseponseModels.ts
@@ -1,7 +1,0 @@
-export interface MuteUserResponse {
-  muteUserOnChannel: boolean;
-}
-
-export interface BanUserResponse {
-  banUserFromChannel: boolean;
-}
